requirements.txt
# Data Processing & Arrays
numpy>=1.24.0
pandas>=2.0.0
xarray>=2023.0.0
netCDF4>=1.6.0

# Scientific & Lidar Specific
atmospheric_lidar>=0.5.0

# Visualization
matplotlib>=3.7.0

# Cloud & Web
boto3>=1.26.0
PyYAML>=6.0